Wolpertinger preset failure fix (#434)
Numpy 1.18 fails to cast float to int as part of the wolpertinger preset run

+++ rl_coach/filters/action/box_discretization.py
@@ -59,7 +59,7 @@ class BoxDiscretization(PartialDiscreteActionSpaceMap):
 
     def get_unfiltered_action_space(self, output_action_space: BoxActionSpace) -> DiscreteActionSpace:
         if isinstance(self.num_bins_per_dimension, int):
-            self.num_bins_per_dimension = np.ones(output_action_space.shape) * self.num_bins_per_dimension
+            self.num_bins_per_dimension = np.ones(output_action_space.shape).astype(int) * self.num_bins_per_dimension
 
         bins = []
         for i in range(len(output_action_space.low)):
